% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesDCG.r
\name{BayesDCG}
\alias{BayesDCG}
\title{Bayesian Cyclic Causal Discovery}
\usage{
BayesDCG(
  data_matrix,
  a_mu,
  b_mu,
  a_gamma,
  b_gamma,
  a_tao,
  b_tao,
  a_gamma_1,
  b_gamma_1,
  alpha,
  M,
  num_iter
)
}
\arguments{
\item{data_matrix}{numeric matrix (N x p) matrix where N is the sample size and p is the number of features}

\item{a_mu}{Hyperparameter for the mean of the normal prior used for each of the cluster means in the mixed normal distribution. Default value is 0.}

\item{b_mu}{Hyperparameter for the variance of the normal prior used for each of the cluster means in the mixed normal distribution. Default value is 2.}

\item{a_gamma}{Hyperparameter for the first shape of the beta prior for the probability there is a node from one feature to another.}

\item{b_gamma}{Hyperparameter for the second shape of the beta prior for the probability there is a node from one feature to another.}

\item{a_tao}{Hyperparameter for the shape of the inverse gamma prior for each of the cluster variance in the mixed normal distribution.}

\item{b_tao}{Hyperparameter for the scale of the inverse gamma prior for each of the cluster variance in the mixed normal distribution.}

\item{alpha}{Hyperparameter for the concentration for the dirichlet prior for the cluster probability of the corresponding feature.}

\item{M}{Number of possible clusters.}

\item{num_iter}{The number of iterations for the bayessclingam algorithm}

\item{a_og_tao}{The shape parameter of the proposal when sampling tao with the adjacency matrix entries}

\item{b_og_tao}{Hyperparameter for the scale of the inverse gamma prior for each of the cluster variance in the mixed normal distribution.}
}
\value{
A list (from the C backend) typically containing adjacency and causal-effect matrices and samples/diagnostics.
}
\description{
An MCMC sampler to generate a directed cyclic graphs (DCG) from provided data and hyperparameters
}
