% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_examples.r
\name{generates_examples}
\alias{generates_examples}
\title{Generating examples}
\usage{
generates_examples(
  num_covariates,
  N,
  M_input,
  prob_sparsity,
  IS_DAG,
  seed_input
)
}
\arguments{
\item{num_covariates}{sample size}

\item{N}{number of features}

\item{prob_sparsity}{Probability when we sample each non diagonal of the Adjacency matrix whether it is 0.}

\item{number}{of clusters for the generated mixed normal}

\item{is_dag}{Whether our true graph is a Directed Acyclic graph or Directed Cyclic graph}
}
\value{
A list containing the data matrix and the true Adjacency matrix
}
\description{
Generates examples to test the two methods provided
}
