% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_est_graph.r
\name{point_est_graph}
\alias{point_est_graph}
\title{Generate summary based on}
\usage{
point_est_graph(
  Adjacency_matrix_list,
  dist_type = "shd",
  dist_fun = NULL,
  burn_in_frac = 0.75
)
}
\arguments{
\item{dist_type}{Character string specifying the distance metric.
One of \code{"shd"}, \code{"sid"}, \code{"forb"}, or \code{"custom"}.}

\item{dist_fun}{Optional user-supplied distance function used when
\code{dist_type = "custom"}. Must have signature \code{dist_fun(A, B)},
where \code{A} and \code{B} are \eqn{p \times p} adjacency matrices and
return a non negative scalar distance.}

\item{burn_in_frac}{Fraction of iterations to discard as burn in (default 0.75).}

\item{posterior_adjacency_matrices}{A matrix of dimension \code{num_iter × (p*p)} where each row is a flattened adjacency matrix sampled during the posterior (0/1 entries).}
}
\value{
The best possible graph structure found through finding the smallest distance through finding the weighted medoid based on chosen distance.
}
\description{
Selects a representative graph structure from posterior samples by computing a \strong{weighted medoid} under a chosen distance. Intended for lists of sampled adjacency matrices
}
\examples{
# Run BayesDAG simulated example

N = 300
num_covariates = 10
M = 2
num_iter = 5000


example_list = generates_examples_DAG(num_covariates, N, M, 0.9, 21)

data_matrix = example_list$data_matrix
Adjacency_matrix_true = example_list$Adjacency_matrix_true

params = list(
a_mu = 0,
b_mu = 2,
a_gamma = 0.5,
b_gamma = 0.5,
a_gamma_1 = 2,
b_gamma_1 = 1,
a_tao = 2,
b_tao = 1,
a_og_tao = 0.01,
b_og_tao = 0.01,
alpha = 1
)


result_list = BayesDAG(
data_matrix,
params$a_mu,
params$b_mu,
params$a_gamma,
params$b_gamma,
params$a_tao,
params$b_tao,
params$a_og_tao,
params$b_og_tao,
params$a_gamma_1,
params$b_gamma_1,
params$alpha,
M,
num_iter
) # Runs the Acyclic algorithm


Adjacency_matrix_list = result_list$Adjacency_matrix_list

Adjacency_matrix = point_est_graph(Adjacency_matrix_list,dist_type = 'shd') # Best graph structure using shd
Adjacency_matrix

Adjacency_matrix = point_est_graph(Adjacency_matrix_list,dist_type = 'sid') # Best graph structure using sid
Adjacency_matrix

custom_edge_mismatch = function(A, B) {
return(sum(abs(A - B)))
}

Adjacency_matrix = point_est_graph(Adjacency_matrix_list, dist_type = 'custom', dist_fun = custom_edge_mismatch)
Adjacency_matrix
}
