% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_adjacency_analysis.r
\name{posterior_adjacency_analysis}
\alias{posterior_adjacency_analysis}
\title{Generate summary based on}
\usage{
posterior_adjacency_analysis(Adjacency_matrix_list, dist_type)
}
\arguments{
\item{dist_type}{Type of distance to use when selecting the medoid (e.g., \code{"shd"}, \code{"sid"},
\code{"forb"}).}

\item{posterior_adjacency_matrices}{A matrix of dimension \code{num_iter × (p*p)} where each row is a flattened adjacency matrix sampled during the posterior (0/1 entries).}
}
\value{
The best possible graph structure found through finding the smallest distance through finding the weighted medoid based on chosen distance.
}
\description{
Selects a representative graph structure from posterior samples by computing a \strong{weighted medoid} under a chosen distance. Intended for lists of sampled adjacency matrices
}
