% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_graph_mass.r
\name{posterior_graph_mass}
\alias{posterior_graph_mass}
\title{Compute Posterior Mass of a Given Graph Structure}
\usage{
posterior_graph_mass(graph_structure, posterior_graph_structures)
}
\arguments{
\item{graph_structure}{An \code{igraph} object representing the target graph.}

\item{posterior_graph_structures}{A matrix of dimension \code{num_iter × p^2}, where each row corresponds
to a flattened adjacency matrix from one MCMC iteration.}
}
\value{
A numeric value between 0 and 1 giving the posterior mass (frequency) with which the target
graph's edges appear in the posterior samples.
}
\description{
This function computes the posterior mass (relative frequency) of a specified graph structure among the collection of
sampled posterior adjacency matrices. The input graph is converted into an adjacency matrix vector, and function
counts how often all edges appear in sampled posterior graph.
}
