% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_interval_est.r
\name{posterior_interval_est}
\alias{posterior_interval_est}
\title{Generate summary based on posterior sample provided}
\usage{
posterior_interval_est(posterior_matrix, level, adjacency = FALSE)
}
\arguments{
\item{posterior_matrix}{numeric posterior sample matrix, where each row corresponds to the MCMC iteration}

\item{level}{Credible level for the intervals, given as a probability between 0 and 1. Only utilized for posterior analysis of non graph structure samples.}
}
\value{
If \code{adjacency = TRUE}, a list with components:
\itemize{
\item \code{pip_matrix:} numeric matrix (P), with edge wise posterior inclusion probabilities, where entry P_{ij} is the proportion of sampled graphs contain an edge from i to j.
\item \code{pip_graph_results:} list in which each element corresponds to a unique graph structure and the proportion of times that structure appears in the posterior graph samples.
}

If \code{adjacency = FALSE}, a list with components:
\itemize{
\item \code{ci_matrix:} numeric matrix, where each row is the equal-tailed credible interval for each parameter provide in the posterior matrix at the
specified level (e.g., 2.5\\% and 97.5\\% for \code{level = 0.95}).
\item \code{hpd_matrix:} numeric matrix, where each row is the HPD interval for each parameter at the same level.
}
}
\description{
Compute posterior summaries from a matrix of MCMC samples. When \code{adjacency = TRUE}, the function treats each row as a flattened
adjacency matrix and returns edge wise posterior inclusion probabilities and posterior probabilities for unique graph structures. Otherwise,
the function treats each column as a parameter and generates a credible and hpd interval based on the level input.
}
\examples{

# Run BayessDAG simulated example

N = 300
num_covariates = 10
M = 2
num_iter = 5000


example_list = generates_examples_DAG(num_covariates, N, M, 0.9, 21)

data_matrix = example_list$data_matrix
Adjacency_matrix_true = example_list$Adjacency_matrix_true

params = list(
a_mu = 0,
b_mu = 2,
a_gamma = 0.5,
b_gamma = 0.5,
a_gamma_1 = 2,
b_gamma_1 = 1,
a_tao = 2,
b_tao = 1,
a_og_tao = 0.01,
b_og_tao = 0.01,
alpha = 1
)


result_list = BayesDAG(
data_matrix,
params$a_mu,
params$b_mu,
params$a_gamma,
params$b_gamma,
params$a_tao,
params$b_tao,
params$a_og_tao,
params$b_og_tao,
params$a_gamma_1,
params$b_gamma_1,
params$alpha,
M,
num_iter
) # Runs the Acyclic algorithm


# Extract posterior outputs
Adjacency_matrix_means = result_list$Adjacency_matrix_means
Adjacency_matrix_list = result_list$Adjacency_matrix_list
Causal_effect_matrix_list = result_list$Causal_effect_matrix_list
gamma_list = result_list$gamma_list
gamma_1_list = result_list$gamma_1_list
mu_matrix_list = result_list$mu_matrix_list
tao_matrix_list = result_list$tao_matrix_list
pi_matrix_list = result_list$pi_matrix_list

Causal_effect_matrix_summary = posterior_interval_est(Causal_effect_matrix_list, level = 0.95)
hpd_matrix_acyclic = Causal_effect_matrix_summary$hpd_matrix
ci_matrix_acyclic = Causal_effect_matrix_summary$ci_matrix

#######################################
# Extracting nonzero HPD intervals
#######################################
par(mfrow=c(2,1))

nonzero_cols = which(colSums(hpd_matrix_acyclic) != 0)
num_non_zero_coef = length(nonzero_cols)
data_1 = data.frame(cbind(1:num_non_zero_coef,t(hpd_matrix_acyclic[,which(colSums(hpd_matrix_acyclic)!=0)])))

nonzero_cols = which(colSums(hpd_matrix_acyclic) != 0)

# subset and transpose so each row = coefficient
hpd_sub = t(hpd_matrix_acyclic[, nonzero_cols, drop = FALSE])
colnames(hpd_sub) = c("lower", "upper")  # row1 = lower, row2 = upper

data_1 = as.data.frame(hpd_sub)
data_1$x = factor(seq_len(nrow(data_1)))

data_1$mid = (data_1$lower + data_1$upper) / 2


ggplot2::ggplot(data_1, ggplot2::aes(x = x, y = mid)) +
 ggplot2::geom_point(size = 3) +
 ggplot2::geom_hline(yintercept = 1, linetype = "dashed", color = "red") +
 ggplot2::geom_errorbar(ggplot2::aes(ymin = lower, ymax = upper), width = 0.2) +
 ggplot2::labs(y = "Causal Weight Estimate with HPD Interval", x = "Nonzero causal effect coefficient (index)") +
 ggplot2::theme_minimal()


#######################################
# Extracting nonzero Credible intervals
#######################################
data_2 = data.frame(ci_matrix_acyclic[which(rowSums(ci_matrix_acyclic)!=0),])
x = factor(1:nrow(data_2))
data_2 = cbind(x,data_2)

ggplot2::ggplot(data_2, ggplot2::aes(x = x, y = X2)) +
  ggplot2::geom_point(size = 3) +
  ggplot2::geom_hline(yintercept = 1, linetype = "dashed", color = "red") +
  ggplot2::geom_errorbar(ggplot2::aes(ymin = X1, ymax = X3), width = 0.2) +  # just X1/X3
  ggplot2::labs(y = "Causal Weight Estimate with 95\% CI", x = "Nonzero causal effect coefficient (index)") +
  ggplot2::theme_minimal()
}
