% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_posterior_matrix.r
\name{summary_posterior_matrix}
\alias{summary_posterior_matrix}
\title{Generate summary based on posterior sample provided}
\usage{
summary_posterior_matrix(posterior_matrix, level, adjacency = FALSE)
}
\arguments{
\item{posterior_matrix}{numeric posterior sample matrix, where each row corresponds to the MCMC iteration}

\item{level}{Credible level for the intervals, given as a probability between 0 and 1. Only utilized for posterior analysis of non graph structure samples.}
}
\value{
If \code{adjacency = TRUE}, a list with components:
\itemize{
\item \code{pip_matrix:} numeric matrix (P), with edge wise posterior inclusion probabilities, where entry P_{ij} is the proportion of sampled graphs contain an edge from i to j.
\item \code{pip_graph_results:} list in which each element corresponds to a unique graph structure and the proportion of times that structure appears in the posterior graph samples.
}

If \code{adjacency = FALSE}, a list with components:
\itemize{
\item \code{ci_matrix:} numeric matrix, where each row is the equal-tailed credible interval for each parameter provide in the posterior matrix at the
specified level (e.g., 2.5\\% and 97.5\\% for \code{level = 0.95}).
\item \code{hpd_matrix:} numeric matrix, where each row is the HPD interval for each parameter at the same level.
}
}
\description{
Compute posterior summaries from a matrix of MCMC samples. When \code{adjacency = TRUE}, the function treats each row as a flattened
adjacency matrix and returns edge wise posterior inclusion probabilities and posterior probabilities for unique graph structures. Otherwise,
the function treats each column as a parameter and generates a credible and hpd interval based on the level input.
}
