% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_examples.r
\name{generates_examples}
\alias{generates_examples}
\title{Generating examples}
\usage{
generates_examples(num_covariates, prob_sparsity, is_dag)
}
\arguments{
\item{num_covariates}{number of features}

\item{prob_sparsity}{Probability when we sample each non diagonal of the Adjacency matrix whether it is 0.}

\item{is_dag}{Whether our true graph is a Directed Acyclic graph or Directed Cyclic graph}
}
\value{
A list containing the data matrix and the true Adjacency matrix
}
\description{
Generates examples to test the two methods provided
}
