% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_examples_DCG.r
\name{generates_examples_DCG}
\alias{generates_examples_DCG}
\title{Generating DCG examples}
\usage{
generates_examples_DCG(num_covariates, N, M_input, prob_sparsity, seed_input)
}
\arguments{
\item{num_covariates}{sample size}

\item{N}{number of features}

\item{prob_sparsity}{Probability when we sample each non diagonal of the Adjacency matrix whether it is 0.}

\item{number}{of clusters for the generated mixed normal}

\item{is_dag}{Whether our true graph is a Directed Acyclic graph or Directed Cyclic graph}
}
\value{
A list containing the data matrix and the true Adjacency matrix
}
\description{
Generates examples to test the two methods provided
}
\examples{
#Run BayesDAG based on simulated example

#Data Generation

set.seed(21)


N = 250 # Sample size of data
num_covariates = 7 # Number of features for test data
M = 2 # Number of finite clusters for mixed normal in likelihood
num_iter = 10000 # Number of iterations MCMC runs

#Function to generate DAG example

example_list = generates_examples_DCG(
num_covariates,
N,
M,
0.9,
21
)

data_matrix = example_list$data_matrix
Adjacency_matrix_true = example_list$Adjacency_matrix_true
Causal_effect_matrix_true = example_list$Causal_effect_matrix_true

#Input parameters for the BayesDAG function

params = list(
 a_mu      = 0,
 b_mu      = 2,
 a_gamma   = 2,
 b_gamma   = 1,
 a_gamma_1 = 2,
 b_gamma_1 = 1,
 a_tao     = 2,
 b_tao     = 1,
 alpha     = 1
)



results_list = BayesDCG(
data_matrix,
params$a_mu,
params$b_mu,
params$a_gamma,
params$b_gamma,
params$a_tao,
params$b_tao,
params$a_gamma_1,
params$b_gamma_1,
params$alpha,
M,
num_iter
)


Adjacency_matrix_means = results_lists$Adjacency_matrix_means
Adjacency_matrix_list = results_lists$Adjacency_matrix_list
Causal_effect_matrix_list = results_lists$Causal_effect_matrix_list
gamma_list = results_lists$gamma_list
gamma_1_list = results_lists$gamma_1_list
mu_matrix_list = results_lists$mu_matrix_list
tao_matrix_list = results_lists$tao_matrix_list
pi_matrix_list = results_lists$pi_matrix_list
}
