% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_posterior_matrix.r
\name{summary_posterior_matrix}
\alias{summary_posterior_matrix}
\title{Generate summary based on posterior sample provided}
\usage{
summary_posterior_matrix(posterior_matrix, level, adjacency = FALSE)
}
\arguments{
\item{posterior_matrix}{numeric posterior sample matrix, where each row corresponds to the MCMC iteration}

\item{level}{Credible level for the intervals, given as a probability between 0 and 1. Only utilized for posterior analysis of non graph structure samples.}
}
\value{
If \code{adjacency = TRUE}, a list with components:
\itemize{
\item \code{pip_matrix:} numeric matrix (P), with edge wise posterior inclusion probabilities, where entry P_{ij} is the proportion of sampled graphs contain an edge from i to j.
\item \code{pip_graph_results:} list in which each element corresponds to a unique graph structure and the proportion of times that structure appears in the posterior graph samples.
}

If \code{adjacency = FALSE}, a list with components:
\itemize{
\item \code{ci_matrix:} numeric matrix, where each row is the equal-tailed credible interval for each parameter provide in the posterior matrix at the
specified level (e.g., 2.5\\% and 97.5\\% for \code{level = 0.95}).
\item \code{hpd_matrix:} numeric matrix, where each row is the HPD interval for each parameter at the same level.
}
}
\description{
Compute posterior summaries from a matrix of MCMC samples. When \code{adjacency = TRUE}, the function treats each row as a flattened
adjacency matrix and returns edge wise posterior inclusion probabilities and posterior probabilities for unique graph structures. Otherwise,
the function treats each column as a parameter and generates a credible and hpd interval based on the level input.
}
\examples{

# Run BayessDAG simulated example

N = 300
num_covariates = 10
M = 2
num_iter = 10000


example_list = generates_examples_DAG(num_covariates, N, M, 0.9, 21)

data_matrix = example_list$data_matrix
Adjacency_matrix_true = example_list$Adjacency_matrix_true

params = list(
a_mu = 0,
b_mu = 2,
a_gamma = 0.5,
b_gamma = 0.5,
a_gamma_1 = 2,
b_gamma_1 = 1,
a_tao = 2,
b_tao = 1,
a_og_tao = 0.01,
b_og_tao = 0.01,
alpha = 1
)


results_lists = BayesDAG(
data_matrix,
params$a_mu,
params$b_mu,
params$a_gamma,
params$b_gamma,
params$a_tao,
params$b_tao,
params$a_og_tao,
params$b_og_tao,
params$a_gamma_1,
params$b_gamma_1,
params$alpha,
M,
num_iter
) # Runs the Acyclic algorithm

}
